Beginning to restructure. Removing relative imports and spiketorch.[...] imports (undoing packaging).
