+++ spiketorch/network/groups.pyc
